Structured Codebase + Added context handling + Added Repo Searching
